## Problem

<!-- What problem are you trying to solve? What issue does this close? -->

Hoping to leverage the concurrent uploads stated in https://docs.aws.amazon.com/cli/latest/topic/s3-config.html to speed up `aws s3 sync`

partially closes https://linear.app/ogp/issue/ISOM-1661/optimise-site-build-times

## Solution

<!-- How did you solve the problem? -->

**Breaking Changes**

<!-- Does this PR contain any backward incompatible changes? If so, what are they and should there be special considerations for release? -->

- [ ] Yes - this PR contains breaking changes
  - Details ...
- [x] No - this PR is backwards compatible

**Improvements**:

- set a higher S3 `max_concurrent_requests` based on the number of cores (arbitrary 4x multiplier since file size is small)
  - We set a floor (20, which is higher than the default 10) and a ceiling (re: https://jasonralph.org/?p=791)
  - experimented with 10x multiplier with no cap (which means 320 concurrency with 32 CPU core, but value is the same). Likely due to resource bottleneck beyond a certain point
- benchmarking
  - before: 140s to 240s
  - after: 80s to 90s 
  - improvement: 50s to 160s reduction